requirements.txt
PySC2==3.0.0
python-dotenv==0.15.0
torch==1.7.0
torchvision==0.8.1
gym==0.17.3
matplotlib==3.3.3
atari-py==0.2.6
